pyproject.toml
[tool.poetry]
name = "self-netdisk"
version = "0.1.0"
description = ""
authors = ["lvtomatoj <lvtomatoj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.16"
fastapi = "^0.110.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
